simpler Outpoint hasher
tested with -reindex-chainstate=1

This didn't improve IBD as much as I thought it would, only about 3%
(real time ratio new/old 0.9722, not counting flushes to disk). But that
could actually be significant over an entire IBD (I only tested to
stopatheight=400000 and dbcache=8000 so there was no flush until the end).