Fix checkpoint failure in MinerTestingSetup::TestBasicMining()
```
$ test_bitcoin --run_test=miner_tests/CreateNewBlock_validity
unknown location(0): fatal error: in "miner_tests/CreateNewBlock_validity": std::runtime_error: CreateNewBlock: TestBlockValidity failed: checkpoint-mismatch
test/miner_tests.cpp(375): last checkpoint
```

This is not an upstream problem. It was introduced in commit 78c54874a5
which added a checkpoint check to ContextualCheckBlockHeader().

The "subsidy changing" part of the failing test creates a chain (of
just the block index entries, not actual blocks) to height 209999, then
attempts to create an actual block (CreateNewBlock()) at next height,
210000. The checkpoint check fails because mainnet happens to have a
checkpoint at that height (and of course the hashes don't match).

We could probably modify the test to not attempt to create a block at
that height, but a better solution is to disable checkpoint checks in
block validation within unit tests. Checkpoint validation in unit tests
never makes sense, because the unit tests use the mainnet chainparams,
and the block hashes will never match the checkpoints.

Fortunately, there's an option to enable checkpoint verification, so
the fix is to set that option to false in the unit test setup.